refactored into prepack react test style
